requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
backcall==0.2.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
customtkinter==5.0.3
darkdetect==0.8.0
debugpy==1.6.5
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.87.0
filetype==1.2.0
flatbuffers==23.1.4
frozenlist==1.3.3
h11==0.14.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.24.0
ipykernel==6.20.1
ipython==8.8.0
jedi==0.18.2
jupyter_client==7.4.9
jupyter_core==5.1.3
llvmlite==0.39.1
matplotlib-inline==0.1.6
mpmath==1.2.1
multidict==6.0.4
mypy-extensions==0.4.3
nest-asyncio==1.5.6
networkx==3.0
numba==0.56.4
numpy==1.23.5
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
packaging==23.0
pandas==1.5.2
parso==0.8.3
pathspec==0.10.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
pooch==1.6.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
pure-eval==0.2.2
pydantic==1.10.4
Pygments==2.14.0
PyMatting==1.1.8
pypdf==3.2.1
PyPDF2==3.0.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7
PyWavelets==1.4.1
pywin32==305
pyzmq==25.0.0
rembg==2.0.30
requests==2.28.2
scikit-image==0.19.3
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.21.0
sympy==1.11.1
tifffile==2022.10.10
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
ttkthemes==3.2.2
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
yarl==1.8.2
